Test for "blob:" in createObjectURL URLs for support
If `URL.createObjectURL` returns a URL that starts with something other
than "blob:", it can be assumed that the `URL` object has been
polyfilled by a library and there is no need to test further.

This addresses the environment created by Blob.js[1] for Internet
Explorer and makes this check for "blob:" URL support fail as it should.

[1]: https://github.com/eligrey/Blob.js/